[clean] refactor [add] add/sub func
