+++ .gitignore
@@ -52,4 +52,4 @@ Mkfile.old
 dkms.conf
 
 tmp*
-slcc$
+slcc
